+++ Changelog.md
@@ -1,3 +1,10 @@
+03/09/2018
+
+* Tested support for Windows 64Bit, MacOSX 32Bit
+* Removed Windows unit from unit Amazon.Utils
+* Added Region to TAmazonClient.Create
+
+
 30/08/2018
 
 * THashSHA2 supported in unit Amazon.Utils for Delphi XE8 and up.